+++ simapp/app_test.go
@@ -106,6 +106,7 @@ func TestRunMigrations(t *testing.T) {
 		name         string
 		moduleName   string
 		fromVersion  uint64
+		toVersion    uint64
 		expRegErr    bool // errors while registering migration
 		expRegErrMsg string
 		expRunErr    bool // errors while running migration
@@ -114,33 +115,33 @@ func TestRunMigrations(t *testing.T) {
 	}{
 		{
 			"cannot register migration for version 0",
-			"bank", 0,
+			"bank", 0, 1,
 			true, "module migration versions should start at 1: invalid version", false, "", 0,
 		},
 		{
 			"throws error on RunMigrations if no migration registered for bank",
-			"", 1,
+			"", 1, 2,
 			false, "", true, "no migrations found for module bank: not found", 0,
 		},
 		{
 			"can register 1->2 migration handler for x/bank, cannot run migration",
-			"bank", 1,
+			"bank", 1, 2,
 			false, "", true, "no migration found for module bank from version 2 to version 3: not found", 0,
 		},
 		{
 			"can register 2->3 migration handler for x/bank, can run migration",
-			"bank", 2,
-			false, "", false, "", 1,
+			"bank", 2, bank.AppModule{}.ConsensusVersion(),
+			false, "", false, "", int(bank.AppModule{}.ConsensusVersion() - 2), // minus 2 because 1-2 is run in the previous test case.
 		},
 		{
 			"cannot register migration handler for same module & fromVersion",
-			"bank", 1,
+			"bank", 1, 2,
 			true, "another migration for module bank and version 1 already exists: internal logic error", false, "", 0,
 		},
 	}
 
 	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
+		t.Run(tc.name, func(tt *testing.T) {
 			var err error
 
 			// Since it's very hard to test actual in-place store migrations in
@@ -150,20 +151,23 @@ func TestRunMigrations(t *testing.T) {
 			called := 0
 
 			if tc.moduleName != "" {
-				// Register migration for module from version `fromVersion` to `fromVersion+1`.
-				err = app.configurator.RegisterMigration(tc.moduleName, tc.fromVersion, func(sdk.Context) error {
-					called++
+				for i := tc.fromVersion; i < tc.toVersion; i++ {
+					// Register migration for module from version `fromVersion` to `fromVersion+1`.
+					tt.Logf("Registering migration for %q v%d", tc.moduleName, i)
+					err = app.configurator.RegisterMigration(tc.moduleName, i, func(sdk.Context) error {
+						called++
 
-					return nil
-				})
+						return nil
+					})
 
-				if tc.expRegErr {
-					require.EqualError(t, err, tc.expRegErrMsg)
+					if tc.expRegErr {
+						require.EqualError(tt, err, tc.expRegErrMsg)
 
-					return
+						return
+					}
+					require.NoError(tt, err, "registering migration")
 				}
 			}
-			require.NoError(t, err)
 
 			// Run migrations only for bank. That's why we put the initial
 			// version for bank as 1, and for all other modules, we put as
@@ -191,11 +195,11 @@ func TestRunMigrations(t *testing.T) {
 				},
 			)
 			if tc.expRunErr {
-				require.EqualError(t, err, tc.expRunErrMsg)
+				require.EqualError(tt, err, tc.expRunErrMsg, "running migration")
 			} else {
-				require.NoError(t, err)
+				require.NoError(tt, err, "running migration")
 				// Make sure bank's migration is called.
-				require.Equal(t, tc.expCalled, called)
+				require.Equal(tt, tc.expCalled, called)
 			}
 		})
 	}